+++ src/style.js
@@ -14,7 +14,8 @@ const GlobalStyles = createGlobalStyle`
   }
 
   html {
-    font-family: sans-serif;
+    font-family: sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 
+      'Segoe UI Symbol', 'Noto Color Emoji';
     box-sizing: border-box;
   }
 
@@ -192,11 +193,24 @@ const Field = styled.div`
   * + * {
     margin-top: 0.375rem;
   }
+
+  & + & {
+    margin-top: 1rem;
+  }
 `;
 
 const Label = styled.label`
   display: block;
   opacity: 0.8;
+`;
+
+const Fieldset = styled.fieldset`
+  margin: 0;
+  border: 1px solid #eee;
+
+  @media (prefers-color-scheme: dark) {
+    border-color: #444;
+  }
 `;
 
 const Input = styled.input`
@@ -213,6 +227,18 @@ const Input = styled.input`
   @media (prefers-color-scheme: dark) {
     background-color: #222;
   }
+`;
+
+const Submit = styled.input`
+  ${normalizeInput}
+
+  border: 0;
+  border-radius: 4px;
+  width: 100%;
+  height: 1.8rem;
+  padding: 0 0.3rem;
+  background-color: ${whatsappThemeColor};
+  color: white;
 `;
 
 const InputDescription = styled.div`
@@ -254,7 +280,9 @@ export {
   SidebarContainer,
   Field,
   Label,
+  Fieldset,
   Input,
+  Submit,
   InputDescription,
   Header,
 };